+++ cas-dotnet-sdk/Hashers/Types/HmacSignByteResult.cs
@@ -5,7 +5,7 @@ namespace CasDotnetSdk.Hashers.Types
     internal struct HmacSignByteResult
     {
         public IntPtr result_bytes_ptr;
-        public int length;
+        public long length;
         public int error_code;
     }
 }
