+++ cas-dotnet-sdk/Hybrid/Types/HpkeEncryptResultStruct.cs
@@ -5,11 +5,11 @@ namespace CasDotnetSdk.Hybrid.Types
     internal struct HpkeEncryptResultStruct
     {
         public IntPtr encapped_key_ptr;
-        public int encapped_key_ptr_length;
+        public long encapped_key_ptr_length;
         public IntPtr ciphertext_ptr;
-        public int ciphertext_ptr_length;
+        public long ciphertext_ptr_length;
         public IntPtr tag_ptr;
-        public int tag_ptr_length;
+        public long tag_ptr_length;
         public int error_code;
     }
 }
