+++ cas-dotnet-sdk/Signatures/Types/Ed25519KeyPairBytesResultStruct.cs
@@ -5,8 +5,8 @@ namespace CasDotnetSdk.Signatures.Types
     internal struct Ed25519KeyPairBytesResultStruct
     {
         public IntPtr signing_key { get; set; }
-        public int signing_key_length { get; set; }
+        public long signing_key_length { get; set; }
         public IntPtr verifying_key { get; set; }
-        public int verifying_key_length { get; set; }
+        public long verifying_key_length { get; set; }
     }
 }
