+++ cas-dotnet-sdk/Sponges/Types/Ascon128DecryptResultStruct.cs
@@ -5,7 +5,7 @@ namespace CasDotnetSdk.Sponges.Types
     internal struct Ascon128DecryptResultStruct
     {
         public IntPtr plaintext { get; set; }
-        public int length { get; set; }
+        public long length { get; set; }
         public int error_code { get; set; }
     }
 }
