+++ cas-dotnet-sdk/Sponges/Types/Ascon128NonceStruct.cs
@@ -5,6 +5,6 @@ namespace CasDotnetSdk.Sponges.Types
     internal struct Ascon128NonceStruct
     {
         public IntPtr nonce;
-        public int length;
+        public long length;
     }
 }
