+++ cas-dotnet-sdk/Symmetric/Types/AesX25519DiffieHellmanStruct.cs
@@ -5,7 +5,7 @@ namespace CasDotnetSdk.Symmetric.Types
     internal struct AesKeyX25519DiffieHellmanStruct
     {
         public IntPtr aes_key_ptr;
-        public int aes_key_ptr_length;
+        public long aes_key_ptr_length;
         public int error_code;
     }
 }
